requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
mlflow>=2.10
dvc>=3.0
pydantic>=2.0
optuna>=3.0
bentoml>=1.2
optuna-integration[mlflow]>=3.0
pyyaml>=6.0
